Progress reorg: You · Ranks · Feed; remove Daily Quests
Collapses the 6-card Progress menu into 3 clear destinations and removes the
"do chores" Daily Quests mechanic (off-brand for a skill/efficiency game).

- Progress submenu → You (/profile) · Ranks (/leaderboard) · Feed (/activity).
- "You" hub (/profile rebuilt): coin avatar + net worth + tabs Stats | History |
  Badges. Stats now matches the public profile — adds Challenge W-L-T, avg & best
  multiple, lifetime earned/spent (was thinner than a stranger's profile).
- Extracted HistoryList + BadgesPanel components, shared by the hub tabs and the
  standalone /history + /badges routes (which now back to "← You").
- get_profile_stats enriched (migration enrich_get_profile_stats; +fix to keep it
  VOLATILE since it calls _ensure_bank which writes).
- Removed Daily Quests: /quests route, getDailyQuests/claimQuestReward wrappers,
  home-page questProgress/refreshQuests + menu badge. (Ledger label kept for history.)

Verified: QA sweep 18/18 pass, 0 console errors across all pages including the new
/profile hub. (Caught + fixed the STABLE-vs-write bug before shipping.)
