+++ src/routes/profile/+page.svelte
@@ -1,79 +1,98 @@
 <script>
   import { onMount } from 'svelte';
   import { goto } from '$app/navigation';
+  import { page } from '$app/stores';
   import { getProfileStats } from '$lib/stores/statsStore.js';
   import { badgeInfo } from '$lib/badges.js';
+  import HistoryList from '$lib/components/HistoryList.svelte';
+  import BadgesPanel from '$lib/components/BadgesPanel.svelte';
   import { track } from '$lib/analytics.js';
 
+  /** @type {'stats'|'history'|'badges'} */
+  let tab = $state('stats');
   /** @type {any|null} */
   let s = $state(null);
   let loading = $state(true);
 
-  onMount(async () => { track('profile_view'); try { s = await getProfileStats(); } finally { loading = false; } });
+  onMount(async () => {
+    track('profile_view');
+    const t = $page.url.searchParams.get('tab');
+    if (t === 'history' || t === 'badges') tab = t;
+    try { s = await getProfileStats(); } finally { loading = false; }
+  });
 
   const fmt = (/** @type {any} */ n) => '$' + Math.round(Number(n ?? 0)).toLocaleString();
+  const mult = (/** @type {any} */ x) => x ? (Number(x) / 100).toFixed(1) + '×' : '—';
   let winPct = $derived(s && s.games_played > 0 ? Math.round((s.games_won / s.games_played) * 100) : 0);
+  let record = $derived(s ? `${s.challenge_wins ?? 0}-${s.challenge_losses ?? 0}-${s.challenge_ties ?? 0}` : '—');
 </script>
 
-<svelte:head><title>WordBank — Profile</title></svelte:head>
+<svelte:head><title>WordBank — You</title></svelte:head>
 
-<main class="profile-page">
+<main class="you-page">
   <button class="back-btn" onclick={() => goto('/')}>← Menu</button>
 
   {#if loading}
     <p class="muted">Loading…</p>
   {:else if s}
-    <h1>{s.username ? '@' + s.username : 'Your Profile'}</h1>
-    <p class="nw-label">Net Worth</p>
-    <div class="nw">{fmt(s.net_worth)}</div>
-    <p class="nw-sub">Your Cash — this is your score</p>
+    <header class="head">
+      <div class="coin" style={s.color ? `--c:${s.color}` : ''}>{(s.username || '?').slice(0, 1).toUpperCase()}</div>
+      <h1>{s.username ? '@' + s.username : 'You'}</h1>
+      <div class="nw" class:neg={Number(s.net_worth) < 0}>{fmt(s.net_worth)}</div>
+      <span class="nw-sub">Net Worth</span>
+    </header>
 
-    <div class="grid">
-      <div class="stat"><span class="sv">🔥 {s.current_streak}</span><span class="sc">Day streak</span></div>
-      <div class="stat"><span class="sv">{s.longest_streak}</span><span class="sc">Longest streak</span></div>
-      <div class="stat"><span class="sv">{winPct}%</span><span class="sc">Daily win rate</span></div>
-      <div class="stat"><span class="sv">{(s.puzzles_solved ?? 0).toLocaleString()}</span><span class="sc">Puzzles solved</span></div>
-      <div class="stat"><span class="sv">#{s.climb_position}</span><span class="sc">Climb position</span></div>
-      <div class="stat"><span class="sv">{s.challenge_wins}</span><span class="sc">Challenge wins</span></div>
+    <div class="tabs">
+      <button class="tab" class:active={tab === 'stats'} onclick={() => tab = 'stats'}>Stats</button>
+      <button class="tab" class:active={tab === 'history'} onclick={() => tab = 'history'}>History</button>
+      <button class="tab" class:active={tab === 'badges'} onclick={() => tab = 'badges'}>Badges</button>
     </div>
 
-    <h2 class="sec">Badges</h2>
-    {#if (s.badges ?? []).length === 0}
-      <p class="muted small">No badges yet — win dailies, climb, and complete weeks to earn them.</p>
-    {:else}
-      <div class="badges">
-        {#each s.badges as id}
-          {@const b = badgeInfo(id)}
-          <div class="badge" title={b.desc}><span class="b-emoji">{b.emoji}</span><span class="b-name">{b.name}</span></div>
-        {/each}
+    {#if tab === 'stats'}
+      <div class="grid">
+        <div class="stat"><span class="sv">🔥 {s.current_streak}</span><span class="sc">Day streak</span></div>
+        <div class="stat"><span class="sv">{s.longest_streak}</span><span class="sc">Best streak</span></div>
+        <div class="stat"><span class="sv">{winPct}%</span><span class="sc">Daily win rate</span></div>
+        <div class="stat"><span class="sv">{(s.puzzles_solved ?? 0).toLocaleString()}</span><span class="sc">Puzzles solved</span></div>
+        <div class="stat"><span class="sv">#{s.climb_position}</span><span class="sc">Climb position</span></div>
+        <div class="stat"><span class="sv">{record}</span><span class="sc">Challenge W-L-T</span></div>
+        <div class="stat"><span class="sv">{mult(s.avg_multiple_x100)}</span><span class="sc">Avg multiple</span></div>
+        <div class="stat"><span class="sv">{mult(s.best_multiple_x100)}</span><span class="sc">Best multiple</span></div>
+        <div class="stat"><span class="sv">{fmt(s.total_earned)}</span><span class="sc">Lifetime earned</span></div>
       </div>
+      <p class="spent-line">Lifetime spent: <b>{fmt(s.total_spent)}</b></p>
+    {:else if tab === 'history'}
+      <HistoryList />
+    {:else}
+      <BadgesPanel />
     {/if}
-
-    <div class="links">
-      <button class="lnk" onclick={() => goto('/badges')}>🏅 All badges</button>
-      <button class="lnk" onclick={() => goto('/leaderboard')}>🏆 Leaderboard</button>
-    </div>
   {/if}
 </main>
 
 <style>
-  .profile-page { max-width: 480px; margin: 0 auto; padding: 1.5rem 1rem 3rem; text-align: center; }
-  .back-btn { display: inline-block; margin-bottom: 1rem; padding: 0.55rem 1.1rem; background: var(--surface); color: var(--text); border: 1px solid var(--border); border-radius: 12px; cursor: pointer; font-weight: 600; font-size: 0.9rem; }
-  .muted { color: var(--text-muted); padding: 1.5rem 0; }
-  .muted.small { font-size: 0.85rem; padding: 0.5rem 0; }
-  h1 { font-family: var(--font-display); font-size: 1.6rem; margin: 0.3rem 0 0.8rem; }
-  .nw-label { color: var(--text-faint); font-size: 0.75rem; text-transform: uppercase; letter-spacing: 0.08em; margin: 0; }
-  .nw { font-family: var(--font-display); font-weight: 800; font-size: 2.6rem; line-height: 1.1; color: var(--brand-2); }
-  .nw-sub { color: var(--text-faint); font-size: 0.82rem; margin: 0.2rem 0 1.4rem; }
-  .grid { display: grid; grid-template-columns: 1fr 1fr 1fr; gap: 0.6rem; }
-  .stat { display: flex; flex-direction: column; gap: 3px; padding: 0.8rem 0.4rem; background: var(--surface); border: 1px solid var(--border); border-radius: 14px; }
-  .sv { font-family: var(--font-display); font-weight: 700; font-size: 1.1rem; color: var(--text); }
+  .you-page { max-width: 520px; margin: 0 auto; padding: 16px 14px 60px; }
+  .back-btn { background: none; border: none; color: var(--text-muted); font-size: 0.92rem; cursor: pointer; padding: 6px 0; }
+  .muted { color: var(--text-muted); text-align: center; padding: 2rem 0; }
+
+  .head { text-align: center; margin: 8px 0 18px; }
+  .coin { width: 64px; height: 64px; margin: 0 auto 8px; border-radius: 50%; display: grid; place-items: center;
+    font-family: var(--font-display); font-weight: 800; font-size: 1.6rem; color: #3a2a00;
+    background: linear-gradient(135deg, var(--c, #fde047), #f59e0b); box-shadow: 0 0 24px rgba(251,191,36,0.4); }
+  h1 { font-family: var(--font-display); font-size: 1.4rem; margin: 0; }
+  .nw { font-family: 'Orbitron', var(--font-display); font-weight: 800; font-size: 2.1rem; color: #fde047;
+    margin-top: 10px; text-shadow: 0 0 18px rgba(251,191,36,0.5); }
+  .nw.neg { color: #fb7185; text-shadow: none; }
+  .nw-sub { font-size: 0.72rem; color: var(--text-faint); text-transform: uppercase; letter-spacing: 0.1em; }
+
+  .tabs { display: flex; gap: 8px; margin: 0 0 16px; }
+  .tab { flex: 1; padding: 9px 0; border-radius: 12px; border: 1px solid var(--border); background: var(--surface);
+    color: var(--text-muted); font-weight: 700; font-size: 0.88rem; cursor: pointer; }
+  .tab.active { background: linear-gradient(135deg, #fde047, #f59e0b); color: #3a2a00; border-color: transparent; }
+
+  .grid { display: grid; grid-template-columns: 1fr 1fr 1fr; gap: 8px; }
+  .stat { display: flex; flex-direction: column; gap: 3px; padding: 0.9rem 0.4rem; background: var(--surface); border: 1px solid var(--border); border-radius: 14px; text-align: center; }
+  .sv { font-family: var(--font-display); font-weight: 800; font-size: 1.05rem; color: var(--text); }
   .sc { font-size: 0.6rem; text-transform: uppercase; letter-spacing: 0.04em; color: var(--text-faint); }
-  .sec { font-family: var(--font-display); font-size: 1.05rem; margin: 1.6rem 0 0.7rem; text-align: left; }
-  .badges { display: flex; flex-wrap: wrap; gap: 0.5rem; }
-  .badge { display: flex; align-items: center; gap: 0.4rem; padding: 0.5rem 0.8rem; background: var(--surface); border: 1px solid rgba(253, 224, 71,0.35); border-radius: 999px; }
-  .b-emoji { font-size: 1rem; }
-  .b-name { font-size: 0.8rem; font-weight: 600; }
-  .links { display: flex; gap: 0.5rem; margin-top: 1.6rem; }
-  .lnk { flex: 1; padding: 0.65rem; border-radius: 12px; border: 1px solid var(--border); background: var(--surface); color: var(--text); cursor: pointer; font-weight: 600; font-size: 0.88rem; }
+  .spent-line { text-align: center; color: var(--text-muted); font-size: 0.84rem; margin: 14px 0 0; }
+  .spent-line b { color: var(--text); }
 </style>
